Add add_relationship abstractmethod
