+++ src/pdag/__init__.py
@@ -2,5 +2,5 @@
 
 __all__ = ["BooleanParameter", "CategoricalParameter", "Model", "NumericParameter", "Relationship", "relationship"]
 from ._decorator import relationship
-from ._model import Model
+from ._model import Model, Relationship
 from ._parameter import BooleanParameter, CategoricalParameter, NumericParameter
